Page Darkens & Small Changes
The main content of the page now darkens when the nav is in its extended
state or the `setDarkened()` function is called with `true`. (I might
need to make a wiki for this project.) Also the nav now has a shadow
which changes on extension and some other small adjustments in the
animations were made.