Implemented the Scroll Link
There is now a new item in the aside (in the nav actually) which scrolls
the page down to the footer and draws some attention to it by darkening
everything except the footer. I used some external libraries for that
but I gave credit to the creators in the code comments.