New Page and a Whole Bunch of Other Things!
Started working on the products page; I also did a lot of restructuring
in the other parts of the project. For example: The CSS is now built in
a more component-oriented way, meaning that elements of the same
category are grouped together in one file. And I did some smaller things
in the footer and the general frame.